server/requirements.txt
Flask
python-dotenv
supabase
pytesseract
opencv-python-headless
Pillow
numpy